Fix Windows compilation

+++ Source/AttoClient/Private/AttoClient.cpp
@@ -21,7 +21,7 @@ FAttoClient::FAttoClient(const FString& Url)
 			return;
 		}
 
-		FBitReader Ar{static_cast<const uint8*>(Data), Size * 8};
+		FBitReader Ar{static_cast<const uint8*>(Data), static_cast<int64>(Size * 8)};
 
 		FAttoS2CProtocol Message;
 		Ar << Message;
